Fixed CreditCardTypeChangeHandler typings
