+++ types/cleave.js/options/creditCard.d.ts
@@ -1,0 +1,19 @@
+import Cleave = require("../");
+
+// Credit Card Options
+export type CreditCardType =
+    | "amex"
+    | "dankort"
+    | "diners"
+    | "discover"
+    | "instapayment"
+    | "jcb"
+    | "maestro"
+    | "mastercard"
+    | "uatp"
+    | "unknown"
+    | "unionPay"
+    | "mir"
+    | "visa";
+
+export type CreditCardTypeChangeHandler = (this: Cleave, type: CreditCardType) => void;
