+++ types/cleave.js/tsconfig.json
@@ -21,7 +21,8 @@
     "files": [
         "cleave.js-tests.tsx",
         "index.d.ts",
-        "options.d.ts",
+        "options/crediCard.d.ts",
+        "options/index.d.ts",
         "react/index.d.ts"
     ]
 }
